TempFS: a (mostly) FakeFS-compatible implementation which uses a real temporary directory
Some code does not work with PyFakeFS (e.g. the sqlite module opens files
in C code which can not be stubbed). Furthermore some tests should actually
access the file system to test Windows-specific behavior.

So TempFS is a class which is mostly interface-compatible and is (almost)
a drop-in replacement for FakeFS. I came to like the use of ".addCleanup"
which simplifies some my traditional "temp directory cleanup" code so this
module will be reused in the future :-)